+++ Runtime/Effects/CellShading/CellShading.shader
@@ -16,10 +16,10 @@
         #include "Packages/com.unity.render-pipelines.core/Runtime/Utilities/Blit.hlsl"
 
         CBUFFER_START(UnityPerMaterial)
-            half4 _BlitTexture_ST;
-            half4 _BlitTexture_TexelSize;
+            float4 _BlitTexture_ST;
+            float4 _BlitTexture_TexelSize;
             int _Shades;
-            half _Strength;
+            float _Strength;
         CBUFFER_END
         ENDHLSL
 
@@ -28,16 +28,16 @@
             HLSLPROGRAM
             #pragma fragment frag
 
-            half4 frag(Varyings input) : SV_TARGET
+            float4 frag(Varyings input) : SV_TARGET
             {
-                half2 uv = input.texcoord;
-                half4 color = SAMPLE_TEXTURE2D(_BlitTexture, sampler_LinearClamp, uv);
+                float2 uv = input.texcoord;
+                float4 color = SAMPLE_TEXTURE2D(_BlitTexture, sampler_LinearClamp, uv);
 
-                half shades = _Shades;
-                half brightness = ((color.r + color.g + color.g) / half(3.0));
-                half shade = floor(brightness * shades);
-                half brighness_of_shade = (shade / shades);
-                half factor = (brightness / brighness_of_shade);
+                float shades = _Shades;
+                float brightness = ((color.r + color.g + color.g) / 3.0);
+                float shade = floor(brightness * shades);
+                float brighness_of_shade = (shade / shades);
+                float factor = (brightness / brighness_of_shade);
                 
                 color.rgb = max(color.rgb * _Strength, color.rgb / factor);
