+++ Runtime/Effects/Fog/Fog.shader
@@ -13,10 +13,10 @@
         #include "Packages/com.unity.render-pipelines.core/Runtime/Utilities/Blit.hlsl"
 
         CBUFFER_START(UnityPerMaterial)
-            half4 _BlitTexture_ST;
-            half4 _BlitTexture_TexelSize;
-            half4 _FogParams;
-            half4 _FogColor;
+            float4 _BlitTexture_ST;
+            float4 _BlitTexture_TexelSize;
+            float4 _FogParams;
+            float4 _FogColor;
         CBUFFER_END
 
         TEXTURE2D_X(_CameraDepthTexture);  
@@ -27,18 +27,18 @@
             HLSLPROGRAM 
             #pragma fragment frag
 
-            half4 frag(Varyings input) : SV_TARGET
+            float4 frag(Varyings input) : SV_TARGET
             {
-                half4 params = _FogParams;
+                float4 params = _FogParams;
                 
-                half2 uv = input.texcoord;
-                half4 color = SAMPLE_TEXTURE2D(_BlitTexture, sampler_LinearClamp, uv);
+                float2 uv = input.texcoord;
+                float4 color = SAMPLE_TEXTURE2D(_BlitTexture, sampler_LinearClamp, uv);
                 
-                half depth = SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture, sampler_LinearClamp, uv);
+                float depth = SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture, sampler_LinearClamp, uv);
                 depth = Linear01Depth(depth, _ZBufferParams);
 
-                half view_distance = (depth * _ProjectionParams.z);
-                half factor = (params.x * max(half(0.0), view_distance - params.y));
+                float view_distance = (depth * _ProjectionParams.z);
+                float factor = (params.x * max(0.0, view_distance - params.y));
                 factor = exp2(-factor * factor);
                 factor = saturate(factor);
